chore: Restructure interface
