+++ src/interfaces/message.interface.ts
@@ -1,0 +1,7 @@
+export interface IMessage {
+  id: string;
+  senderId: string;
+  recipientId: string;
+  text: string;
+  createdAt: any;
+}
